Removed return value of jump position from CountJump#CountJump() and CountJump#JumpFunc(); it isn't needed, as these functions are typically used directly in motion mappings
CountJump#JumpFunc() now uses cursor position after invoking jump function, and doesn't require a returned position any more. This is only a special case for CountJump#TextObject, and should not be generally required of a jump function. The jump function is now also expected to beep, so removed that here.