Rename g:CountJump_Context to g:CountJump_TextObjectContext
When using CountJump#Region#TextObject#Make(), the generated jump functions also set that context, and there's a clash with CountJump#TextObject#TextObjectWithJumpFunctions().
In particular, one cannot store a context for the entire text object (both jumps to begin and end), as the individual jump functions clear the identical context.
Clear g:CountJump_TextObjectContext at the end of the function.